pyproject.toml
[tool.poetry]
name = "Kbot Vector Demo"
version = "0.1.0"
description = ""
authors = ["Walter <jinshuhaicc@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.23.0"
oracledb = "^2.5.0"
python-decouple = "^3.8"
requests = "^2.32.3"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
